Move macros into internal package
This cleans up the import when doing `import prometheus4cats._`
