+++ core/src/main/scala-2/prometheus4cats/internal/MacroUtils.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package prometheus4cats
+package prometheus4cats.internal
 
 import scala.reflect.macros.blackbox
