+++ core/src/main/scala-3/prometheus4cats/internal/CounterNameFromStringLiteral.scala
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-package prometheus4cats
+package prometheus4cats.internal
+
+import prometheus4cats._
 
 import scala.quoted.*
