+++ core/src/main/scala-3/prometheus4cats/internal/LabelNameFromStringLiteral.scala
@@ -14,37 +14,46 @@
  * limitations under the License.
  */
 
-package prometheus4cats
+package prometheus4cats.internal
+
+import prometheus4cats._
 
 import scala.quoted.*
 
-trait InfoNameFromStringLiteral {
+trait LabelNameFromStringLiteral {
 
-  inline def apply(inline t: String): Info.Name = ${
-    InfoNameFromStringLiteral.nameLiteral('t)
+  inline def apply(inline t: String): Label.Name = ${
+    LabelNameFromStringLiteral.nameLiteral('t)
   }
 
-  implicit inline def fromStringLiteral(inline t: String): Info.Name = ${
-    InfoNameFromStringLiteral.nameLiteral('t)
+  implicit inline def fromStringLiteral(inline t: String): Label.Name = ${
+    LabelNameFromStringLiteral.nameLiteral('t)
   }
 
 }
 
-object InfoNameFromStringLiteral extends MacroUtils {
-  def nameLiteral(s: Expr[String])(using q: Quotes): Expr[Info.Name] =
+object LabelNameFromStringLiteral extends MacroUtils {
+  def nameLiteral(s: Expr[String])(using q: Quotes): Expr[Label.Name] =
     s.value match {
       case Some(string) =>
-        Info.Name
+        Label.Name
           .from(string)
           .fold(
             error,
             _ =>
               '{
-                Info.Name.from(${ Expr(string) }).toOption.get
+                Label.Name
+                  .from(${
+                    Expr(string)
+                  })
+                  .toOption
+                  .get
               }
           )
       case None =>
-        abort("Info.Name.from")
-        '{ ??? }
+        abort("Label.Name.from")
+        '{
+          ???
+        }
     }
 }
