+++ core/src/main/scala-3/prometheus4cats/internal/MacroUtils.scala
